requirements.txt
playwright==1.41.0
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
